+++ CHANGELOG.md
@@ -23,6 +23,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `party` condition now supports a location variable, to be used in a static context
 - `enitity` condition now supports variables for the entity name
 - `mmochangeclass` objective for MMOCore that listens for a player changing their class
+- TrainCarts integration
+  - `traincartslocation` objective that executes when the player passed a location 
+  - `traincartsride` objective that executes when the player is riding a train
+  - `traincartsexit` objective that executes when the player exited a train
+  - `traincartsride` condition to check if the player is riding a specific named train
 ### Changed
 - `spawn` event now only spawn mobs and no other entities
 ### Deprecated